added target y and fit the model
